reporter plugin: Move AbstractReporterImpl under org.embulk.spi
