+++ embulk-core/src/main/java/org/embulk/exec/NullReporterPlugin.java
@@ -4,7 +4,7 @@ import org.embulk.config.ConfigSource;
 import org.embulk.config.Task;
 import org.embulk.config.TaskSource;
 import org.embulk.spi.ReporterPlugin;
-import org.embulk.spi.util.AbstractReporterImpl;
+import org.embulk.spi.AbstractReporterImpl;
 
 import javax.annotation.concurrent.ThreadSafe;
