+++ embulk-core/src/main/java/org/embulk/spi/AbstractReporterImpl.java
@@ -1,4 +1,4 @@
-package org.embulk.spi.util;
+package org.embulk.spi;
 
 import com.google.common.collect.ImmutableMap;
 import org.embulk.spi.Reporter;
