+++ embulk-core/src/main/java/org/embulk/spi/ExecSession.java
@@ -1,6 +1,5 @@
 package org.embulk.spi;
 
-import org.embulk.spi.util.AbstractReporterImpl;
 import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.ILoggerFactory;
